+++ Firebase/Auth/Source/RPCs/FIRVerifyAssertionRequest.m
@@ -120,9 +120,9 @@ static NSString *const kSessionIDKey = @"sessionId";
                                                       value:_providerAccessToken]];
   }
 
-  if (!_providerIDToken && !_providerAccessToken && !_requestURI) {
+  if (!_providerIDToken && !_providerAccessToken && !_pendingToken && !_requestURI) {
     [NSException raise:NSInvalidArgumentException
-                format:@"Either IDToken or accessToken must be supplied."];
+        format:@"One of IDToken, accessToken, pendingToken, or requestURI must be supplied."];
   }
 
   if (_providerOAuthTokenSecret) {