+++ Firestore/Example/Tests/API/FSTAPIHelpers.mm
@@ -95,8 +95,8 @@ FIRQuerySnapshot *FSTTestQuerySnapshot(
     const absl::string_view path,
     NSDictionary<NSString *, NSDictionary<NSString *, id> *> *oldDocs,
     NSDictionary<NSString *, NSDictionary<NSString *, id> *> *docsToAdd,
-    BOOL hasPendingWrites,
-    BOOL fromCache) {
+    bool hasPendingWrites,
+    bool fromCache) {
   FIRSnapshotMetadata *metadata =
       [FIRSnapshotMetadata snapshotMetadataWithPendingWrites:hasPendingWrites fromCache:fromCache];
   FSTDocumentSet *oldDocuments = FSTTestDocSet(FSTDocumentComparatorByKey, @[]);