+++ GoogleUtilities/Logger/GULASLLogger.m
@@ -119,9 +119,8 @@ NS_ASSUME_NONNULL_BEGIN
 }
 
 - (void)logWithLevel:(GULLoggerLevel)level
-         withService:(GULLoggerService)service
+         withService:(GULLoggerService)__unused service
             isForced:(BOOL)forced
-            withCode:(NSString *)messageCode
          withMessage:(NSString *)message {
   // Skip logging this if the level isn't to be logged unless it's forced.
   if (![self isLoggableLevel:level] && !forced) {