+++ Firebase/Core/CHANGELOG.md
@@ -3,6 +3,11 @@
   Please call the new APIs directly: Enable/disable Analytics with `Analytics.setAnalyticsEnabled(_)`
   and modify the session timeout interval with `Analytics.setAnalyticsCollectionEnabled(_)`.
 
+# 2019-03-19 -- v5.4.0 -- M45
+- [changed] Allow Bundle IDs that have a valid prefix to enable richer extension support. (#2515)
+- [changed] Convert FIRLogger to use the recommended os_log for logging via GoogleUtilities.
+  See https://developer.apple.com/documentation/os/logging?language=objc.
+
 # 2019-01-22 -- v5.2.0 -- M41
 - [changed] Added a registerInternalLibrary API. Now other Firebase libraries register with FirebaseCore
   instead of FirebaseCore needing all of its clients' versions built in.